fix path, rename widgetCardChart to chart
